Use thread_local as hart local

+++ kernel/inc/k_main.h
@@ -9,7 +9,7 @@ extern "C"
 
     int k_boot(void **);
     int k_boot_perip();
-    unsigned long k_boot_harts(int);
+    int k_boot_harts(int);
 
     int k_premain(int hartid);
     int k_main(int hartid);
